Split CGTownInstance file
