+++ lib/mapObjects/MapObjects.h
@@ -15,8 +15,10 @@
 
 #include "CArmedInstance.h"
 #include "CBank.h"
+#include "CGDwelling.h"
 #include "CGHeroInstance.h"
 #include "CGMarket.h"
+#include "CGTownBuilding.h"
 #include "CGTownInstance.h"
 #include "CGPandoraBox.h"
 #include "CRewardableObject.h"
